Set timezone to +00:00 to avoid problems there
There can be problems because of dst when storing timestamps. It can be read about here: http://stackoverflow.com/a/1650910/147949
But it seems to be solved by setting timezone to +00:00. I Guess if the first hour was set on source-db using from_timestamp then this still can produce inconstencies that probably best be solved by selecting and inserting using unixtimestamp-functions