Better handling of timestamps
Now dealing with timestamps in their integer representation; FROM_UNIXTIME() and UNIX_TIMESTAMP().
This is to avoid problems with leaphours. Because when selecting and stringifying a timestamp which is on  a leaphour is lossy. There's no way to know from the string if it represents the hour before or after the leap. When inserted and turned into a unix-integer, the hour after the leap is assumed.
As a plus, dealing with timestamps this way should also improve performance.